Add more sofia tests
Add a full sofia section rendering test as well as a new aliases test
using the new sequence API.

Additionally,
- move config templates into the conftest
- use default sofia profile stop timeouts
