+++ tests/test_directory.py
@@ -8,43 +8,6 @@ import pytest
 
 
 @pytest.fixture
-def domain_template(request, confmng):
-    """Template for a directory domain in Python dict form based on
-    the default config.
-    """
-    dialstring = '{^^:sip_invite_domain=${dialed_domain}:presence_id=${dialed_user}@${dialed_domain}}${sofia_contact(*/${dialed_user}@${dialed_domain})}'
-    pw = 'doggypants'
-
-    return {
-        'params': {'dial-string': dialstring},
-        'variables': {
-            "record_stereo": "true",
-            "default_gateway": "$${default_provider}",
-            "default_areacode": "$${default_areacode}",
-            "transfer_fallback_extension": "operator",
-        },
-        'groups': {'default': {'users': {
-            'doggy': {
-                'params': {
-                    'password': pw,
-                    'vm-password': pw,
-                },
-                'variables': {
-                    "toll_allow": "domestic,international,local",
-                    "accountcode": "1000",
-                    "user_context": "default",
-                    "effective_caller_id_name": "Extension 1000",
-                    "effective_caller_id_number": "1000",
-                    "outbound_caller_id_name": "$${outbound_caller_name}",
-                    "outbound_caller_id_number": "$${outbound_caller_id}",
-                    "callgroup": "techsupport",
-                }
-            }
-        }}},
-    }
-
-
-@pytest.fixture
 def domain(confmng, domain_template):
     name = 'test'  # '{}'.format(confmng.cli.eval('$${local_ip_v4}'))
     confmng.directory[name] = domain_template
